% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_cases_heat}
\alias{age_cases_heat}
\title{Historical Weekly COVID-19 Cases by Age for Indiana}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 592 rows and 6 columns.
}
\usage{
age_cases_heat
}
\description{
Processed data for Demographics \href{https://ercbk.github.io/Indiana-COVID-19-Website/demographics.html#cases-by-age}{heatmaps} that shows a breakdown of weekly COVID-19 cases by age group.
}
\details{
Data Dictionary \href{https://ercbk.github.io/indianacovid19data/articles/data-dictionary.html#historical-weekly-covid-19-cases-by-age-for-indiana-1}{entry}
}
\examples{
head(age_cases_heat)
}
\keyword{datasets}
