% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_death_line}
\alias{age_death_line}
\title{Historical Weekly COVID-19 Deaths by Age Group for Indiana}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 680 rows and 5 columns.
}
\usage{
age_death_line
}
\description{
Processed data for Demographics \href{https://ercbk.github.io/Indiana-COVID-19-Website/demographics.html#deaths-by-age}{line charts} that shows counts and trends of weekly COVID-19 deaths by age group.
}
\details{
Data Dictionary \href{https://ercbk.github.io/indianacovid19data/articles/data-dictionary.html#historical-weekly-covid-19-deaths-by-age-group-for-indiana-1}{entry}
}
\examples{
head(age_death_line)
}
\keyword{datasets}
