% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_hosp_line}
\alias{age_hosp_line}
\title{Historical Daily COVID-19 Hospital Admissions by Age Group for Indiana}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 873 rows and 5 columns.
}
\usage{
age_hosp_line
}
\description{
Raw data for the Hospitals \href{https://ercbk.github.io/Indiana-COVID-19-Website/hospitals.html#state-hospital-mortality-staffing-shortages-admissions}{line charts} that shows daily COVID-19 hospital admissions by age group and gender.
}
\details{
Data Dictionary \href{https://ercbk.github.io/indianacovid19data/articles/data-dictionary.html#historical-daily-covid-19-hospital-admissions-by-age-group-for-indiana-1}{entry}
}
\examples{
head(age_hosp_line)
}
\keyword{datasets}
