% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beds_vents_complete}
\alias{beds_vents_complete}
\title{Historical Daily ICU Beds and Ventilators for Indiana}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 302 rows and 8 columns.
}
\usage{
beds_vents_complete
}
\description{
Raw data for the Static Charts \href{https://ercbk.github.io/Indiana-COVID-19-Website/static.html#Hospitalizations,_ICU_Beds_and_Ventilator_Availability}{Hospitalizations, ICU Beds and Ventilator Availability} that shows resource counts of ICU beds and ventilators.
}
\details{
Data Dictionary \href{https://ercbk.github.io/indianacovid19data/articles/data-dictionary.html#historical-daily-icu-beds-and-ventilators-for-indiana-1}{entry}
}
\examples{
head(beds_vents_complete)
}
\keyword{datasets}
