% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dash_ci_line}
\alias{dash_ci_line}
\title{Historical Daily Values of Delphi Research Groups Combined Indicator for Indiana}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4425 rows and 3 columns.
}
\usage{
dash_ci_line
}
\description{
Processed data for the \href{https://ercbk.github.io/Indiana-COVIDcast-Dashboard/#dashboard}{Indiana COVIDcast Dashboard} that shows historic daily values of Delphi Research Group's Combined Indicator for each methropolitan statistical area.
}
\details{
Data Dictionary \href{https://ercbk.github.io/indianacovid19data/articles/data-dictionary.html#historical-daily-values-of-delphi-research-groups-combined-indicator-for-indiana-1}{entry}
}
\examples{
head(dash_ci_line)
}
\keyword{datasets}
