% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goog_mob_ind}
\alias{goog_mob_ind}
\title{Historical Google Maps Mobility Indices for Indiana}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1692 rows and 4 columns.
}
\usage{
goog_mob_ind
}
\description{
Processed data for the Google Maps Mobility Indices \href{https://ercbk.github.io/Indiana-COVID-19-Website/static.html#Google_Maps_Mobility_Indices}{line chart} that shows percent differences from a mobility baseline that was calculated during the 5-week period Jan 3 – Feb 6, 2020. These percent changes show how lengths-of-stay at different locations have changed from pre-pandemic conditions.
}
\details{
Data Dictionary \href{https://ercbk.github.io/indianacovid19data/articles/data-dictionary.html#historical-google-maps-mobility-indices-for-indiana-1}{entry}
}
\examples{
head(goog_mob_ind)
}
\keyword{datasets}
