% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hosp_msas_line}
\alias{hosp_msas_line}
\title{Historical State Hospital Staff Shortages, Hospital Mortality Rate, Hospital Admissions and Ages Skewness of Admissions for Indiana}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 95 rows and 5 columns.
}
\usage{
hosp_msas_line
}
\description{
Processed data for Hospitals \href{https://ercbk.github.io/Indiana-COVID-19-Website/hospitals.html#state-hospital-mortality-staffing-shortages-admissions}{line charts} that shows state hospital staff shortages, hospital mortality rate, hospital admissions, and age skewness of hospital admissions towards older patients.
}
\details{
Data Dictionary \href{https://ercbk.github.io/IndianaCOVID19TrackerData/articles/data-dictionary.html#historical-state-hospital-staff-shortages-hospital-mortality-rate-hospital-admissions-and-ages-skewness-of-admissions-for-indiana-1}{entry}
}
\examples{
head(hosp_msas_line)
}
\keyword{datasets}
