% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hosp_react_tab}
\alias{hosp_react_tab}
\title{Local Hospital Capacity for Indiana}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 125 rows and 10 columns.
}
\usage{
hosp_react_tab
}
\description{
Processed data for the Hospitals \href{https://ercbk.github.io/Indiana-COVID-19-Website/hospitals.html#local-hospital-capacity}{Local Hospital Capacity table} that shows capacity measures and counts of COVID-19 patients for hospitals at the local level.
}
\details{
Data Dictionary \href{https://ercbk.github.io/indianacovid19data/articles/data-dictionary.html#local-hospital-capacity-for-indiana-1}{entry}
}
\examples{
head(hosp_react_tab)
}
\keyword{datasets}
