% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{illinois_tests_complete}
\alias{illinois_tests_complete}
\title{Historical Weekly COVID-19 Statistics and Counts for Illinois Counties}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 5768 rows and 10 columns.
}
\usage{
illinois_tests_complete
}
\description{
Raw data for the \href{https://ercbk.github.io/Indiana-COVIDcast-Dashboard/#dashboard}{Indiana COVIDcast Dashboard} that shows various weekly COVID-19 county statistics for Illinois
}
\details{
Data Dictionary \href{https://ercbk.github.io/indianacovid19data/articles/data-dictionary.html#historical-weekly-covid-19-statistics-and-counts-for-illinois-counties-1}{entry}
}
\examples{
head(illinois_tests_complete)
}
\keyword{datasets}
