% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ind_age_complete}
\alias{ind_age_complete}
\title{Historical Daily Tests, Cases and Deaths by Age for Indiana}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3302 rows and 8 columns.
}
\usage{
ind_age_complete
}
\description{
Raw data for Demographics \href{https://ercbk.github.io/Indiana-COVID-19-Website/demographics.html}{line charts} that shows daily case and death counts by age group.
}
\details{
Data Dictionary \href{https://ercbk.github.io/indianacovid19data/articles/data-dictionary.html#historical-daily-tests-cases-and-deaths-by-age-for-indiana-1}{entry}
}
\examples{
head(ind_age_complete)
}
\keyword{datasets}
