% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ind_race_complete}
\alias{ind_race_complete}
\title{Historical Daily Cases and Deaths by Race for Indiana}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4160 rows and 19 columns.
}
\usage{
ind_race_complete
}
\description{
Raw data that shows daily case and death counts by race.
}
\examples{
head(ind_race_complete)
}
\keyword{datasets}
