% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{median_age_bubble}
\alias{median_age_bubble}
\title{Historical Weekly Median Age of Cases, Weekly Tests and Deaths for Indiana}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51 rows and 4 columns.
}
\usage{
median_age_bubble
}
\description{
Processed data for the Demographics \href{https://ercbk.github.io/Indiana-COVID-19-Website/demographics.html#median-age-of-cases-tests-deaths}{bubble chart} that shows the number of tests, number of deaths, and median age of the cases per week.
}
\details{
Data Dictionary \href{https://ercbk.github.io/indianacovid19data/articles/data-dictionary.html#historical-weekly-median-age-of-cases-weekly-tests-and-deaths-for-indiana-1}{entry}
}
\examples{
head(median_age_bubble)
}
\keyword{datasets}
