% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mich_tests_complete}
\alias{mich_tests_complete}
\title{Historical Daily COVID-19 Test Results for Michigan Counties}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 44922 rows and 5 columns.
}
\usage{
mich_tests_complete
}
\description{
Raw data for the \href{https://ercbk.github.io/Indiana-COVIDcast-Dashboard/#dashboard}{Indiana COVIDcast Dashboard} that shows daily COVID-19 county test results for Michigan
}
\details{
Data Dictionary \href{https://ercbk.github.io/indianacovid19data/articles/data-dictionary.html#historical-daily-covid-19-test-results-for-michigan-counties-1}{entry}
}
\examples{
head(mich_tests_complete)
}
\keyword{datasets}
