% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mort_hosp_line}
\alias{mort_hosp_line}
\title{Historical Daily Hospital Admissions and Deaths for Indiana}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 205 rows and 4 columns.
}
\usage{
mort_hosp_line
}
\description{
Raw data for the Hospitals \href{https://ercbk.github.io/Indiana-COVID-19-Website/hospitals.html#state-hospital-mortality-staffing-shortages-admissions}{line charts} that shows daily COVID-19 hospital admissions, hospital deaths, and hospital death rate.
}
\details{
Data Dictionary \href{https://ercbk.github.io/indianacovid19data/articles/data-dictionary.html#historical-daily-hospital-admissions-and-deaths-for-indiana-1}{entry}
}
\examples{
head(mort_hosp_line)
}
\keyword{datasets}
