% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msa_cases100_posrate_historic}
\alias{msa_cases100_posrate_historic}
\title{Historical Weekly Positivity Rates and Daily Cases per 100,000 Indiana Residents}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 5145 rows and 7 columns.
}
\usage{
msa_cases100_posrate_historic
}
\description{
Processed data for the \href{https://ercbk.github.io/Indiana-COVIDcast-Dashboard/#dashboard}{Indiana COVIDcast Dashboard} that shows weekly positivity rate and daily cases per 100,000 residents per metropolitan statistical area.
}
\examples{
head(msa_cases100_posrate_historic)
}
\keyword{datasets}
