% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{open_tab_reg}
\alias{open_tab_reg}
\title{OpenTable Year-Over-Year Seated Diners for Regional States}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3732 rows and 4 columns.
}
\usage{
open_tab_reg
}
\description{
Processed data for the OpenTable Year-Over-Year Seated Diners \href{https://ercbk.github.io/Indiana-COVID-19-Website/static.html#OpenTable_Year-Over-Year_Seated_Diners_}{line charts} that shows reservation data from a sample of restaurants across Indiana and other regional states. Each data point is the median daily percent difference in seated diners from 2019 to either 2020 or 2021 depending on the date.
}
\details{
Data Dictionary \href{https://ercbk.github.io/indianacovid19data/articles/data-dictionary.html#opentable-year-over-year-seated-diners-for-regional-states-1}{entry}
}
\examples{
head(open_tab_reg)
}
\keyword{datasets}
