% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-to-file.R
\name{path_to_file}
\alias{path_to_file}
\title{Get file path to .csv files}
\source{
The function is similar to \code{\link[palmerpenguins:path_to_file]{palmerpenguins::path_to_file()}} with a few edits.
}
\usage{
path_to_file(path = NULL)
}
\arguments{
\item{path}{Name of file in quotes with extension. If \code{NULL}, the available files will be listed.}
}
\description{
indianacovid19data has csv file types for all its datasets in the \code{inst/extdata}
directory. This function makes them easy to access.
}
\details{
The csv file for \link{hosp_react_tab} doesn't include the list columns.
}
\examples{
path_to_file()
path_to_file("hosp_msas_line.csv")
head(read.csv(path_to_file("hosp_msas_line.csv")))
}
