% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ind_race_complete}
\alias{ind_race_complete}
\title{Historical Daily Cases and Deaths by Race for Indiana}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4160 rows and 58 columns.
}
\usage{
ind_race_complete
}
\description{
Raw data that shows daily case and death counts by race.
}
\details{
Data Dictionary \href{https://ercbk.github.io/indianacovid19data/articles/data-dictionary.html#historical-daily-cases-and-deaths-by-race-for-indiana-1}{entry}
}
\examples{
head(ind_race_complete)
}
\keyword{datasets}
