% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ind_tests_complete}
\alias{ind_tests_complete}
\title{Historical Daily COVID-19 Test Results for Indiana Counties}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 51060 rows and 4 columns.
}
\usage{
ind_tests_complete
}
\description{
Raw data for the \href{https://ercbk.github.io/Indiana-COVIDcast-Dashboard/#dashboard}{Indiana COVIDcast Dashboard} that shows daily COVID-19 county test results for Indiana
}
\details{
Data Dictionary \href{https://ercbk.github.io/indianacovid19data/articles/data-dictionary.html#historical-daily-covid-19-test-results-for-indiana-counties-1}{entry}
}
\examples{
head(ind_tests_complete)
}
\keyword{datasets}
